Region Time
Region Time is not an app. nor a website, Region Time Is a file On your PC Witch has a Path. and you can open that Path in your Browser to run Code in it by Opening the File. You can also Save it to your Browser like its a real website. But its not. and The path is different for everyone. Some may be same. But remember this isn't a site. Its literally a file that tells you the time when you open its Path on your Browser!